Moved components into a /components folder, renamed articles-container to articles, and updated app.js accordingly.
